Add MurmurHash3 (unused) and misc cleanup
Added the ability to use the MurmurHash3 algorithm (but you'd have to
change some comments around to actually enable it).

Removed some duplicated lines in src/bloomfilter.h (were already
imported from src/mmapbitarray.h).

Fixed up schizophrenic type returned by _filesize(int fd).  It was
trying to return an int (-1 indicating error) and getting stored in a
uint64_t which obviously cannot even represent a -1.  So now the value
0xffffffffffffffff is used to indicate failure.

Rebuilt pybloomfilter.c with Cython 0.17.

+++ src/mmapbitarray.h
@@ -63,7 +63,7 @@ MBArray * mbarray_Copy_Template(MBArray * src, char * filename, int perms);
 int mbarray_Update(MBArray * array, char * data, int size);
 /*MBArray * mbarray_Copy(MBarray * src, const char * filename);*/
 
-int mbarray_FileSize(MBArray * array);
+uint64_t mbarray_FileSize(MBArray * array);
 
 char * mbarray_CharData(MBArray * array);
